requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
easyocr==1.7.1
pillow==10.4.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.8.2
python-multipart==0.0.9